feat: Base UI Improvement in Browser Extension
- Fix config for style in Browser Extension build
- Update env.example

Resolves #10
