+++ app/.env.example
@@ -1,1 +1,1 @@
-EXPO_PUBLIC_API_URL=
+EXPO_PUBLIC_API_URL=http://your-ip-address:5001/api
