+++ backend/.env.example
@@ -1,3 +1,3 @@
-CONNECTION_URL=
+CONNECTION_URL=mongodb-connection-url
 PORT=5001
-JWT_SECRET=
+JWT_SECRET=appsecret
