+++ extension/.env.example
@@ -1,1 +1,1 @@
-REACT_APP_API_URL=
+REACT_APP_API_URL=http://localhost:5001/api
