+++ extension/package.json
@@ -2,7 +2,7 @@
   "name": "browsewipe",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "react/index.jsx",
   "type": "module",
   "license": "AGPL-version-3.0",
   "private": false,
@@ -24,22 +24,30 @@
   },
   "contributors": [],
   "scripts": {
-    "dev": "",
+    "dev": "webpack --config webpack.config.js --mode development --watch",
     "test": "",
     "build": "webpack --config webpack.config.js --mode production"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.28.3",
     "@babel/preset-react": "^7.27.1",
+    "@tailwindcss/postcss": "^4.1.13",
+    "autoprefixer": "^10.4.21",
     "babel-loader": "^10.0.0",
     "copy-webpack-plugin": "^13.0.1",
+    "css-loader": "^7.1.2",
     "html-webpack-plugin": "^5.6.4",
+    "mini-css-extract-plugin": "^2.9.4",
+    "postcss": "^8.5.6",
+    "postcss-loader": "^8.2.0",
+    "style-loader": "^4.0.0",
     "webpack": "^5.101.3",
     "webpack-cli": "^6.0.1"
   },
   "dependencies": {
     "dotenv": "^17.2.2",
     "react": "^19.1.1",
-    "react-dom": "^19.1.1"
+    "react-dom": "^19.1.1",
+    "tailwindcss": "^3.4.17"
   }
 }
