+++ extension/tailwind.config.js
@@ -1,0 +1,11 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: [
+    "./src/**/*.{js,jsx,ts,tsx}",
+    "./src/react/index.html"
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
